refactor: organize merchant route tests by scope
