+++ backend/tests/routes/merchants/__init__.py
@@ -1,0 +1,1 @@
+"""Merchant route tests"""
